# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Scalable Railway Ticketing Platform Authentication Contract (Scoped)
  version: 1.1.0
  description: >-
    This document intentionally covers only the public registration and login
    contract. It is not a complete specification of the platform API.
paths:
  /api/v1/auth/register:
    post:
      summary: Request customer account registration
      description: >-
        A syntactically valid registration request receives the same response
        whether or not the normalized email address is already registered.
        Registration never returns access or refresh credentials; the customer
        must use the login endpoint after the account exists.
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "202":
          description: Registration request accepted without disclosing account existence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationAccepted"
              example:
                message: If the registration request can be processed, the account workflow will continue.
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/Unavailable"
  /api/v1/auth/login:
    post:
      summary: Authenticate an existing account
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Authentication succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          description: Credentials were not accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
        "503":
          $ref: "#/components/responses/Unavailable"
components:
  schemas:
    RegisterRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
        - display_name
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 12
        display_name:
          type: string
          minLength: 1
    LoginRequest:
      type: object
      additionalProperties: false
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 1
    RegistrationAccepted:
      type: object
      additionalProperties: false
      required:
        - message
      properties:
        message:
          type: string
          const: If the registration request can be processed, the account workflow will continue.
    TokenPair:
      type: object
      additionalProperties: false
      required:
        - access_token
        - refresh_token
        - token_type
        - expires_in
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          const: Bearer
        expires_in:
          type: integer
          format: int64
          minimum: 1
    ErrorEnvelope:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string
  responses:
    InvalidRequest:
      description: The request is invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Unavailable:
      description: Service is temporarily unavailable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"